chore: switched to unit testing
